atualizando o frontend com mais melhorias nas telas e filtros de date picker nas movimentaçõesfinanceiras, tambem adicionei no sidebar as opcoes de cadastros e configuracoes (ainda sem crud). No backend, fz a parte de inativacao de categoria financeira e categoria de produtos, assim como toda a logica de nao criar titulos ou produtos com uma categoria inativada e a parte de exibicao das mesmas que sera tratado de forma diferente no frontend

+++ backend/src/main/java/com/estoqueplan/estoque_plan/service/CategoriaService.java
@@ -2,8 +2,11 @@ package com.estoqueplan.estoque_plan.service;
 
 import com.estoqueplan.estoque_plan.model.Categoria;
 import com.estoqueplan.estoque_plan.repository.CategoriaRepository;
+import com.estoqueplan.estoque_plan.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,20 +14,46 @@ import java.util.Optional;
 public class CategoriaService {
     @Autowired
     private CategoriaRepository categoriaRepository;
+    private ProdutoRepository produtoRepository;
 
     public Categoria salvarCategoria(Categoria categoria) {
         return categoriaRepository.save(categoria);
     }
 
-    public List<Categoria> listarCategorias() {
-        return categoriaRepository.findAll();
+    public List<Categoria> listarCategorias(Boolean incluirInativos) {
+        if (Boolean.TRUE.equals(incluirInativos)) {
+            return categoriaRepository.findAll();
+        }
+        return categoriaRepository.findByAtivoTrue();
     }
 
     public Optional<Categoria> buscarPorId(Long id) {
         return categoriaRepository.findById(id);
     }
 
+    public void inativarCategoriaPorId(Long id) {
+        Categoria categoria = categoriaRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Categoria não encontrada!"));
+
+        if (!categoria.isAtivo()) {
+            return;
+        }
+
+        categoria.setAtivo(false);
+        categoria.setInativadoEm(LocalDateTime.now());
+        categoriaRepository.save(categoria);
+    }
+
     public void deletarCategoria(Long id) {
+        if (!categoriaRepository.existsById(id)) {
+            throw new RuntimeException("Categoria não encontrada!");
+        }
+
+        boolean temProdutoAssoaciado = produtoRepository.existsByCategoriaId(id);
+        if (temProdutoAssoaciado) {
+            throw new RuntimeException("Não é possível excluir: existem produtos vinculados a esta categoria!");
+        }
+
         categoriaRepository.deleteById(id);
     }
 }
